Fix unhandled rejection errors for test based on redis in feathers. Remove mockery for this tests..
