+++ src/utils/redis.js
@@ -6,10 +6,10 @@ const logger = require('../logger');
 
 let redisClient = false;
 
-async function initializeRedisClient() {
+async function initializeRedisClient(optionalRedisInstance) {
 	if (Configuration.has('REDIS_URI')) {
 		try {
-			redisClient = new Redis(Configuration.get('REDIS_URI'));
+			redisClient = optionalRedisInstance || new Redis(Configuration.get('REDIS_URI'));
 
 			// The error event must be handled, otherwise the app crashes on redis connection errors.
 			// This is due to basic NodeJS behavior: https://nodejs.org/api/events.html#error-events
