+++ test/services/account/services/jwtTimerService.test.js
@@ -1,59 +1,79 @@
-const assert = require('assert');
-
 const { expect } = require('chai');
-const mockery = require('mockery');
 const commons = require('@hpi-schul-cloud/commons');
 
 const { Configuration } = commons;
 const whitelist = require('../../../../src/services/authentication/logic/whitelist');
-const redisMock = require('../../../utils/redis/redisMock');
 const appPromise = require('../../../../src/app');
-const testObjects = require('../../helpers/testObjects')(appPromise());
+const testHelper = require('../../helpers/testObjects');
+const redisHelper = require('../../../../src/utils/redis');
 
 const { setupNestServices, closeNestServices } = require('../../../utils/setup.nest.services');
 
+const valueDict = {};
+const ttlDict = {};
+class RedisClientMock {
+	get(key) {
+		const value = valueDict[key];
+		return Promise.resolve(value);
+	}
+
+	set(key, value, ...args) {
+		valueDict[key] = value;
+		const ex = args.indexOf('EX');
+		if (ex >= 0) {
+			ttlDict[key] = args[ex + 1];
+		}
+		return Promise.resolve(true);
+	}
+
+	del(key) {
+		delete valueDict[key];
+		delete ttlDict[key];
+		return Promise.resolve(true);
+	}
+
+	ttl(key) {
+		const ttl = ttlDict[key];
+		return Promise.resolve(ttl);
+	}
+
+	on(_key, _func) {}
+}
+
 describe('jwtTimer service', () => {
 	describe('redis mocked', () => {
-		let app;
-		let nestServices;
-		let redisHelper;
-		let configBefore = null;
-
 		describe('with redis instance', () => {
+			let app;
 			let server;
+			let configBefore = null;
+			let nestServices;
+			let testObjects;
+
 			before(async () => {
 				app = await appPromise();
+				testObjects = testHelper(app);
 				server = await app.listen(0);
 				nestServices = await setupNestServices(app);
 
 				configBefore = Configuration.toObject({ plainSecrets: true });
-				mockery.enable({
-					warnOnReplace: false,
-					warnOnUnregistered: false,
-					useCleanCache: true,
-				});
-				delete require.cache[require.resolve('../../../../src/utils/redis')];
-				delete require.cache[require.resolve('../../../../src/services/account/services/jwtTimerService')];
 
-				mockery.registerMock('ioredis', redisMock);
-				mockery.registerMock('@hpi-schul-cloud/commons', commons);
-				/* eslint-disable global-require */
-				redisHelper = require('../../../../src/utils/redis');
+				// eslint-disable-next-line global-require
 				const { jwtTimerServiceSetup } = require('../../../../src/services/account/services/jwtTimerService');
 				app.unuse('/accounts/jwtTimer');
 				app.configure(jwtTimerServiceSetup);
 				/* eslint-enable global-require */
 
+				// we need clean file that no redis instance is saved from any test before
+				delete require.cache[require.resolve('../../../../src/utils/redis')];
 				Configuration.set('REDIS_URI', '//validHost:4444');
-				redisHelper.initializeRedisClient();
+				const redisMock = new RedisClientMock();
+				redisHelper.initializeRedisClient(redisMock);
 			});
 
 			after(async () => {
-				mockery.deregisterAll();
-				mockery.disable();
 				await testObjects.cleanup();
-				delete require.cache[require.resolve('../../../../src/utils/redis')];
 				delete require.cache[require.resolve('../../../../src/services/account/services/jwtTimerService')];
+				delete require.cache[require.resolve('../../../../src/utils/redis')];
 				Configuration.reset(configBefore);
 				await server.close();
 				await closeNestServices(nestServices);
@@ -84,34 +104,31 @@ describe('jwtTimer service', () => {
 
 		describe('without redis instance', () => {
 			let server;
+			let app;
+			let nestServices;
+			let testObjects;
 
 			before(async () => {
 				app = await appPromise();
+				testObjects = testHelper(app);
 				server = await app.listen(0);
 				nestServices = await setupNestServices(app);
 
-				mockery.enable({
-					warnOnReplace: false,
-					warnOnUnregistered: false,
-					useCleanCache: true,
-				});
-				mockery.registerMock('ioredis', redisMock);
-
+				// we need clean file that no redis instance is saved from any test before
 				delete require.cache[require.resolve('../../../../src/utils/redis')];
+
 				/* eslint-disable global-require */
-				redisHelper = require('../../../../src/utils/redis');
 				const { jwtTimerServiceSetup } = require('../../../../src/services/account/services/jwtTimerService');
 				app.unuse('/accounts/jwtTimer');
 				app.configure(jwtTimerServiceSetup);
-				/* eslint-enable global-require */
 
 				redisHelper.initializeRedisClient();
 			});
 
 			after(async () => {
-				mockery.deregisterAll();
-				mockery.disable();
 				await testObjects.cleanup();
+				delete require.cache[require.resolve('../../../../src/services/account/services/jwtTimerService')];
+				delete require.cache[require.resolve('../../../../src/utils/redis')];
 				await server.close();
 				await closeNestServices(nestServices);
 			});
@@ -121,9 +138,9 @@ describe('jwtTimer service', () => {
 				const params = await testObjects.generateRequestParamsFromUser(user);
 				try {
 					await app.service('/accounts/jwtTimer').find(params);
-					throw new Error('should have failed');
+					throw new Error('Should not fail!');
 				} catch (err) {
-					expect(err.message).to.not.equal('should have failed');
+					expect(err.message).to.not.equal('Should not fail!');
 					expect(err.code).to.equal(405);
 					expect(err.message).to.equal('This feature is disabled on this instance!');
 				}
@@ -134,21 +151,13 @@ describe('jwtTimer service', () => {
 				const params = await testObjects.generateRequestParamsFromUser(user);
 				try {
 					await app.service('/accounts/jwtTimer').create({}, params);
-					throw new Error('should have failed');
+					throw new Error('Should not fail!');
 				} catch (err) {
-					expect(err.message).to.not.equal('should have failed');
+					expect(err.message).to.not.equal('Should not fail!');
 					expect(err.code).to.equal(405);
 					expect(err.message).to.equal('This feature is disabled on this instance!');
 				}
 			});
-		});
-
-		after(async () => {
-			await testObjects.cleanup();
-
-			delete require.cache[require.resolve('../../../../src/utils/redis')];
-			delete require.cache[require.resolve('../../../../src/services/account/services/jwtTimerService')];
-			Configuration.reset(configBefore);
 		});
 	});
 });
