--- test/utils/redis/redisMock.js
@@ -1,31 +1,0 @@
-const valueDict = {};
-const ttlDict = {};
-
-const RedisClientMock = class {
-	get(key) {
-		const value = valueDict[key];
-		return Promise.resolve(value);
-	}
-
-	set(key, value, ...args) {
-		valueDict[key] = value;
-		const ex = args.indexOf('EX');
-		if (ex >= 0) {
-			ttlDict[key] = args[ex + 1];
-		}
-		return Promise.resolve(true);
-	}
-
-	del(key) {
-		delete valueDict[key];
-		delete ttlDict[key];
-		return Promise.resolve(true);
-	}
-
-	ttl(key) {
-		const ttl = ttlDict[key];
-		return Promise.resolve(ttl);
-	}
-};
-
-module.exports = RedisClientMock;
